pyproject.toml
[project]
name = "mindframe"
version = "0.1.0"
description = "Video analysis platform — transcripts, key frames, and content pattern analysis"
requires-python = ">=3.10"

dependencies = [
    "youtube-transcript-api>=1.0.0",
    "yt-dlp>=2024.0.0",
    "click>=8.0.0",
    "streamlit>=1.30.0",
    "requests>=2.31.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=5.0.0",
]

[project.scripts]
mindframe = "src.cli:main"

[tool.pytest.ini_options]
testpaths = ["tests"]
